PICSURE-208 separate mail and event listening, store some variables
