+++ pic-sure-auth-services/src/main/resources/accessEmail.mustache
@@ -4,8 +4,8 @@ You have been granted access to {{system}} with the following roles and privileg
 {{#roles}}
 {{name}} : {{#privileges}} {{description}} {{/privileges}}
 {{/roles}}
-You can find documentation at the following links:
  {{#documentation}}
+You can find documentation at the following links:
  {{.}}
 {{/documentation}}
